Split `pie serve` subcommand into a separate module.
